feat: add cy.provider() for an injectable eip1193 (#11)
Adds a cy.provider() command which yields an EIP-1193 provider to be
injected into an AUTWindow.

Adds unit tests and a cypress test (to check Cypress-specific typings
and as a sanity test). Includes general cleanup, as this work uncovered
a few other issues: eg resetting the chain between tests not modifying
state was very time intensive, so it is now skipped.